feat: added new field to account for deactivation, added queryset, reorganize and split admin , adapt account admin
