+++ iaso/admin/bulk_create_user_csv_file.py
@@ -1,0 +1,8 @@
+from django.contrib.gis import admin
+
+from iaso.models import BulkCreateUserFile
+
+
+@admin.register(BulkCreateUserFile)
+class BulkCreateUserFile(admin.ModelAdmin):
+    pass
