+++ iaso/admin/external_credentials.py
@@ -1,0 +1,8 @@
+from django.contrib.gis import admin
+
+from iaso.models import ExternalCredentials
+
+
+@admin.register(ExternalCredentials)
+class ExternalCredentialsAdmin(admin.ModelAdmin):
+    pass
