+++ iaso/admin/org_unit_change_request_configuration.py
@@ -1,0 +1,8 @@
+from django.contrib.gis import admin
+
+from iaso.models import OrgUnitChangeRequestConfiguration
+
+
+@admin.register(OrgUnitChangeRequestConfiguration)
+class OrgUnitChangeRequestConfigurationAdmin(admin.ModelAdmin):
+    autocomplete_fields = ["project"]
