+++ iaso/admin/reports.py
@@ -1,0 +1,13 @@
+from django.contrib.gis import admin
+
+from iaso.models import Report, ReportVersion
+
+
+@admin.register(Report)
+class ReportAdmin(admin.ModelAdmin):
+    pass
+
+
+@admin.register(ReportVersion)
+class ReportVersionAdmin(admin.ModelAdmin):
+    pass
