+++ iaso/admin/user_roles.py
@@ -1,0 +1,8 @@
+from django.contrib.gis import admin
+
+from iaso.models import UserRole
+
+
+@admin.register(UserRole)
+class UserRoleAdmin(admin.ModelAdmin):
+    autocomplete_fields = ["account"]
