+++ plugins/registry/admin.py
@@ -1,7 +1,7 @@
 from django.contrib.gis import admin
 from django.db import models
 
-from iaso.admin import IasoJSONEditorWidget
+from iaso.admin.utils import IasoJSONEditorWidget
 
 from .models import PublicRegistryConfig
